requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
tqdm
faiss-cpu
scikit-learn
numpy
pydantic
gigachat
typer[all]
python-dotenv